Added a label property to DataSet, and a factory method to create label data batches for training.
