+++ Sources/MimicComputeEngine/Types/Operational.swift
